+++ src/forge_io/readers/__init__.py
@@ -1,7 +1,8 @@
-"""Format readers (OIIO, future vendor SDKs)."""
+"""Format readers (OIIO, ffmpeg, vendor SDKs)."""
 
 from forge_io.readers.arri_reader import ArriRawReader
+from forge_io.readers.ffmpeg_reader import FFmpegReader
 from forge_io.readers.oiio_reader import OIIOReader
 from forge_io.readers.red_reader import RedRawReader
 
-__all__ = ["ArriRawReader", "OIIOReader", "RedRawReader"]
+__all__ = ["ArriRawReader", "FFmpegReader", "OIIOReader", "RedRawReader"]